# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareForm AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patient/{id}:
    get:
      operationId: getPatient
      tags: [patient]
      summary: Get patient intake
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient intake data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updatePatient
      tags: [patient]
      summary: Update patient intake
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "200":
          description: Updated patient intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patient/{id}/pass:
    get:
      operationId: getActivePass
      tags: [patient]
      summary: Get the current active (non-expired, non-revoked) pass for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivePass"
        "404":
          description: No active pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patient/{id}/shared-with:
    get:
      operationId: getSharedWith
      tags: [patient]
      summary: Get the full list of providers this patient has shared their CarePass with
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared-with list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedPassList"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patient/{id}/shared-with/{code}/revoke:
    delete:
      operationId: revokeSharedPass
      tags: [patient]
      summary: Revoke a specific shared pass by code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Pass revoked
        "404":
          description: Shared pass not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patient/{id}/access-history:
    get:
      operationId: getAccessHistory
      tags: [patient]
      summary: Get the access history for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Access history entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessHistoryList"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /codes:
    post:
      operationId: createCode
      tags: [codes]
      summary: Generate a 6-digit access code for a patient (revokes any existing active code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
      responses:
        "201":
          description: Access code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivePass"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /codes/{code}:
    get:
      operationId: validateCode
      tags: [codes]
      summary: Validate a 6-digit access code and return patient intake (logs the view)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient intake data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeValidationResult"
        "404":
          description: Code not found, expired, or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: revokeCode
      tags: [codes]
      summary: Revoke an active access code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Code revoked
        "404":
          description: Code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        errorCode:
          type: string
      required:
        - error
        - errorCode
    Allergy:
      type: object
      properties:
        name:
          type: string
        reaction:
          type: string
        severity:
          type: string
      required:
        - name
        - reaction
        - severity
    Medication:
      type: object
      properties:
        name:
          type: string
        dose:
          type: string
        frequency:
          type: string
        route:
          type: string
        prescriber:
          type: string
        reason:
          type: string
      required:
        - name
        - dose
        - frequency
        - route
        - prescriber
        - reason
    Condition:
      type: object
      properties:
        name:
          type: string
        diagnosedDate:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - name
        - diagnosedDate
        - status
        - notes
    Surgery:
      type: object
      properties:
        procedure:
          type: string
        date:
          type: ["string", "null"]
        facility:
          type: ["string", "null"]
      required:
        - procedure
        - date
        - facility
    Hospitalization:
      type: object
      properties:
        reason:
          type: string
        date:
          type: string
        facility:
          type: string
      required:
        - reason
        - date
        - facility
    Immunization:
      type: object
      properties:
        vaccine:
          type: string
        date:
          type: string
      required:
        - vaccine
        - date
    FamilyHistoryEntry:
      type: object
      properties:
        relation:
          type: string
        condition:
          type: string
      required:
        - relation
        - condition
    SocialHistory:
      type: object
      properties:
        smoking:
          type: string
        alcohol:
          type: string
        occupation:
          type: string
        employer:
          type: string
        exercise:
          type: string
      required:
        - smoking
        - alcohol
        - occupation
        - employer
        - exercise
    CareTeam:
      type: object
      properties:
        pcp:
          type: string
        referringPhysician:
          type: string
        visitSpecialty:
          type: string
        reasonForVisit:
          type: string
        preferredPharmacy:
          type: string
        pharmacyAddress:
          type: string
        pharmacyPhone:
          type: string
      required:
        - pcp
        - referringPhysician
        - visitSpecialty
        - reasonForVisit
        - preferredPharmacy
        - pharmacyAddress
        - pharmacyPhone
    Insurance:
      type: object
      properties:
        plan:
          type: string
        memberId:
          type: string
        group:
          type: string
        policyholder:
          type: string
        policyholderDob:
          type: string
        policyholderRelationship:
          type: string
        phone:
          type: string
      required:
        - plan
        - memberId
        - group
        - policyholder
        - policyholderDob
        - policyholderRelationship
        - phone
    EmergencyContact:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        phone:
          type: string
      required:
        - name
        - relationship
        - phone
    ResponsibleParty:
      type: object
      properties:
        name:
          type: string
        relationship:
          type: string
        dob:
          type: string
        phone:
          type: string
        address:
          type: string
        employer:
          type: string
      required:
        - name
        - relationship
        - dob
        - phone
        - address
        - employer
    Vitals:
      type: object
      properties:
        heightFt:
          type: string
        heightIn:
          type: string
        weightLbs:
          type: string
        systolic:
          type: string
        diastolic:
          type: string
      required:
        - heightFt
        - heightIn
        - weightLbs
        - systolic
        - diastolic
...
    SharedPassList:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/SharedPass"
      required:
        - entries
...
    Patient:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        biologicalSex:
          type: string
        genderIdentity:
          type: string
        preferredName:
          type: string
        pronouns:
          type: string
        preferredLanguage:
          type: string
        maritalStatus:
          type: string
        bloodType:
          type: string
        ssnLastFour:
          type: string
        race:
          type: string
        ethnicity:
          type: string
        interpreterNeeded:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        careTeam:
          $ref: "#/components/schemas/CareTeam"
        insurance:
          $ref: "#/components/schemas/Insurance"
        insuranceSecondary:
          oneOf:
            - $ref: "#/components/schemas/Insurance"
            - type: "null"
        responsibleParty:
          oneOf:
            - $ref: "#/components/schemas/ResponsibleParty"
            - type: "null"
        emergencyContact:
          $ref: "#/components/schemas/EmergencyContact"
        allergies:
          type: array
          items:
            $ref: "#/components/schemas/Allergy"
        medications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/Condition"
        surgeries:
          type: array
          items:
            $ref: "#/components/schemas/Surgery"
        hospitalizations:
          type: array
          items:
            $ref: "#/components/schemas/Hospitalization"
        immunizations:
          type: array
          items:
            $ref: "#/components/schemas/Immunization"
        familyHistory:
          type: array
          items:
            $ref: "#/components/schemas/FamilyHistoryEntry"
        socialHistory:
          $ref: "#/components/schemas/SocialHistory"
        vitals:
          $ref: "#/components/schemas/Vitals"
        reviewOfSystems:
          $ref: "#/components/schemas/ReviewOfSystems"
        obgynHistory:
          oneOf:
            - $ref: "#/components/schemas/ObgynHistory"
            - type: "null"
        consents:
          $ref: "#/components/schemas/Consents"
        signature:
          oneOf:
            - $ref: "#/components/schemas/PatientSignature"
            - type: "null"
        recentUpdates:
          type: array
          items:
            $ref: "#/components/schemas/RecentUpdate"
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - dateOfBirth
        - biologicalSex
        - genderIdentity
        - preferredName
        - pronouns
        - preferredLanguage
        - maritalStatus
        - bloodType
        - ssnLastFour
        - race
        - ethnicity
        - interpreterNeeded
        - phone
        - email
        - address
        - careTeam
        - insurance
        - insuranceSecondary
        - responsibleParty
        - emergencyContact
        - allergies
        - medications
        - conditions
        - surgeries
        - hospitalizations
        - immunizations
        - familyHistory
        - socialHistory
        - vitals
        - reviewOfSystems
        - obgynHistory
        - consents
        - signature
        - recentUpdates
        - updatedAt
    PatientInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        biologicalSex:
          type: string
        genderIdentity:
          type: string
        preferredName:
          type: string
        pronouns:
          type: string
        preferredLanguage:
          type: string
        maritalStatus:
          type: string
        bloodType:
          type: string
        ssnLastFour:
          type: string
        race:
          type: string
        ethnicity:
          type: string
        interpreterNeeded:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        careTeam:
          $ref: "#/components/schemas/CareTeam"
        insurance:
          $ref: "#/components/schemas/Insurance"
        insuranceSecondary:
          oneOf:
            - $ref: "#/components/schemas/Insurance"
            - type: "null"
        responsibleParty:
          oneOf:
            - $ref: "#/components/schemas/ResponsibleParty"
            - type: "null"
        emergencyContact:
          $ref: "#/components/schemas/EmergencyContact"
        allergies:
          type: array
          items:
            $ref: "#/components/schemas/Allergy"
        medications:
          type: array
          items:
            $ref: "#/components/schemas/Medication"
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/Condition"
        surgeries:
          type: array
          items:
            $ref: "#/components/schemas/Surgery"
        hospitalizations:
          type: array
          items:
            $ref: "#/components/schemas/Hospitalization"
        immunizations:
          type: array
          items:
            $ref: "#/components/schemas/Immunization"
        familyHistory:
          type: array
          items:
            $ref: "#/components/schemas/FamilyHistoryEntry"
        socialHistory:
          $ref: "#/components/schemas/SocialHistory"
        vitals:
          $ref: "#/components/schemas/Vitals"
        reviewOfSystems:
          $ref: "#/components/schemas/ReviewOfSystems"
        obgynHistory:
          oneOf:
            - $ref: "#/components/schemas/ObgynHistory"
            - type: "null"
        consents:
          $ref: "#/components/schemas/Consents"
        signature:
          oneOf:
            - $ref: "#/components/schemas/PatientSignature"
            - type: "null"
        recentUpdates:
          type: array
          items:
            $ref: "#/components/schemas/RecentUpdate"
      required:
        - firstName
        - lastName
        - dateOfBirth
        - biologicalSex
        - genderIdentity
        - preferredName
        - pronouns
        - preferredLanguage
        - maritalStatus
        - bloodType
        - ssnLastFour
        - race
        - ethnicity
        - interpreterNeeded
        - phone
        - email
        - address
        - careTeam
        - insurance
        - insuranceSecondary
        - responsibleParty
        - emergencyContact
        - allergies
        - medications
        - conditions
        - surgeries
        - hospitalizations
        - immunizations
        - familyHistory
        - socialHistory
        - vitals
        - reviewOfSystems
        - obgynHistory
        - consents
        - signature
    CodeRequest:
      type: object
      properties:
        patientId:
          type: string
      required:
        - patientId
    ActivePass:
      type: object
      properties:
        code:
          type: string
        patientId:
          type: string
        expiresAt:
          type: string
        revokedAt:
          type: ["string", "null"]
      required:
        - code
        - patientId
        - expiresAt
        - revokedAt
...
    AccessHistoryList:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AccessHistoryEntry"
      required:
        - entries
    CodeValidationResult:
      type: object
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        expiresAt:
          type: string
        lastViewedAt:
          type: ["string", "null"]
      required:
        - patient
        - expiresAt
        - lastViewedAt